project/requirements.txt
django==3.2.7
djangorestframework==3.13.1
django-cors-headers==3.13.0
gunicorn
django-environ==0.8.1
psycopg2
redis==4.3.1
celery==5.2.6
